Retire server.ts and use index.ts
